[PLAYER-5446] Some entities renamed to match structure
