+++ sdk/react/src/lib/collapser.js
@@ -6,7 +6,7 @@ const constants = {
   keep: 'keep'
 };
 
-const CollapsingBarUtils = {
+const Collapser = {
   // @param barWidth numeric.
   // @param orderedItems array of left to right ordered items. Each item meets the skin's 'button' schema.
   // @return {fit:[items that fit in the barWidth], overflow:[items that did not fit]}.
@@ -95,4 +95,4 @@ const CollapsingBarUtils = {
   }
 };
 
-module.exports = CollapsingBarUtils;
+module.exports = Collapser;
